modifiche varie

+++ src/main/java/com/project/Health_Bot/model/Alimento.java
@@ -60,7 +60,7 @@ public class Alimento {
      * 
      * @return energia
      */
-    public double getKcal() {
+    public float getKcal() {
         return kcal;
     }
